temporarly removed audio/subtitles to refactor

+++ setup.py
@@ -11,48 +11,50 @@ from setuptools import (
 from Cython.Build import cythonize
 from Cython.Compiler.AutoDocTransforms import EmbedSignature
 
-IS_PLATFORM_DARWIN = bool(sys.platform == "darwin")
-IS_PLATFORM_DARWIN_ARM = IS_PLATFORM_DARWIN and bool(platform.machine() == "arm64")
-IS_PLATFORM_WINDOWS = bool(sys.platform == "win32")
-IS_PLATFORM_LINUX = bool(sys.platform == "linux")
+rebuilt: bool = False
+if not rebuilt:
+    IS_PLATFORM_DARWIN = bool(sys.platform == "darwin")
+    IS_PLATFORM_DARWIN_ARM = IS_PLATFORM_DARWIN and bool(platform.machine() == "arm64")
+    IS_PLATFORM_WINDOWS = bool(sys.platform == "win32")
+    IS_PLATFORM_LINUX = bool(sys.platform == "linux")
 
-def get_platform_tag() -> str:
-    """return platform tag"""
-    machine = platform.machine()
-    if IS_PLATFORM_LINUX:
-        return f"manylinux_{machine}"
+    def get_platform_tag() -> str:
+        """return platform tag"""
+        machine = platform.machine()
+        if IS_PLATFORM_LINUX:
+            return f"manylinux_{machine}"
 
-    elif IS_PLATFORM_WINDOWS:
-        if struct.calcsize("P") * 8 == 64:
-            return "win_amd64"
+        elif IS_PLATFORM_WINDOWS:
+            if struct.calcsize("P") * 8 == 64:
+                return "win_amd64"
+            else:
+                return "win32"
         else:
-            return "win32"
-    else:
-        raise Exception(f"Unsupported platform {sys.platform}")
+            raise Exception(f"Unsupported platform {sys.platform}")
 
 
 
-ffmpeg_workdir = f"build/tmp/deploy/ffmpeg-{get_platform_tag()}"
-ffmpeg_lib_dir = os.path.join(ffmpeg_workdir, "lib")
-ffmpeg_include_dir = os.path.join(ffmpeg_workdir, "include")
+    ffmpeg_workdir = f"build/tmp/deploy/ffmpeg-{get_platform_tag()}"
+    ffmpeg_lib_dir = os.path.join(ffmpeg_workdir, "lib")
+    ffmpeg_include_dir = os.path.join(ffmpeg_workdir, "include")
 
-package_dir = "src/pylibav"
-package_lib_dir = os.path.join(package_dir, "lib")
-package_include_dir = os.path.join(package_dir, "include")
+    package_dir = "src/pylibav"
+    package_lib_dir = os.path.join(package_dir, "lib")
+    package_include_dir = os.path.join(package_dir, "include")
 
-# Copy FFmpeg AV libraries to package source
-if os.path.exists(package_lib_dir):
-    shutil.rmtree(package_lib_dir)
-os.makedirs(package_lib_dir)
+    # Copy FFmpeg AV libraries to package source
+    if os.path.exists(package_lib_dir):
+        shutil.rmtree(package_lib_dir)
+    os.makedirs(package_lib_dir)
 
-lib_pattern = f"[^.]+.so.*" if IS_PLATFORM_LINUX else f"[^.]+.dll"
-for f in os.listdir(ffmpeg_lib_dir):
-    fp: str = os.path.join(ffmpeg_lib_dir, f)
-    if (
-        not os.path.islink(fp)
-        and (match := re.match(re.compile(lib_pattern), f))
-    ):
-        shutil.copy2(fp, package_lib_dir)
+    lib_pattern = f"[^.]+.so.*" if IS_PLATFORM_LINUX else f"[^.]+.dll"
+    for f in os.listdir(ffmpeg_lib_dir):
+        fp: str = os.path.join(ffmpeg_lib_dir, f)
+        if (
+            not os.path.islink(fp)
+            and (match := re.match(re.compile(lib_pattern), f))
+        ):
+            shutil.copy2(fp, package_lib_dir)
 
 
 FFMPEG_AV_LIBRARIES = [
@@ -76,25 +78,25 @@ print()
 
 
 
-# Monkey-patch Cython to not overwrite embedded signatures.
-old_embed_signature = EmbedSignature._embed_signature
+# # Monkey-patch Cython to not overwrite embedded signatures.
+# old_embed_signature = EmbedSignature._embed_signature
 
-def new_embed_signature(self, sig, doc):
-    # Strip any `self` parameters from the front.
-    sig = re.sub(r"\(self(,\s+)?", "(", sig)
+# def new_embed_signature(self, sig, doc):
+#     # Strip any `self` parameters from the front.
+#     sig = re.sub(r"\(self(,\s+)?", "(", sig)
 
-    # If they both start with the same signature; skip it.
-    if sig and doc:
-        new_name = sig.split("(")[0].strip()
-        old_name = doc.split("(")[0].strip()
-        if new_name == old_name:
-            return doc
-        if new_name.endswith("." + old_name):
-            return doc
+#     # If they both start with the same signature; skip it.
+#     if sig and doc:
+#         new_name = sig.split("(")[0].strip()
+#         old_name = doc.split("(")[0].strip()
+#         if new_name == old_name:
+#             return doc
+#         if new_name.endswith("." + old_name):
+#             return doc
 
-    return old_embed_signature(self, sig, doc)
+#     return old_embed_signature(self, sig, doc)
 
-EmbedSignature._embed_signature = new_embed_signature
+# EmbedSignature._embed_signature = new_embed_signature
 
 
 
@@ -113,6 +115,9 @@ for dirname, dirnames, filenames in os.walk(package_dir):
             os.path.splitext(pyx_filepath)[0]
         ).replace("/", ".").replace(os.sep, ".")
 
+        if "filtergraph" in module_name:
+            continue
+
         # Cythonize the module.
         ext_modules.extend(
             cythonize(
@@ -126,8 +131,8 @@ for dirname, dirnames, filenames in os.walk(package_dir):
                 compiler_directives=dict(
                     c_string_type="str",
                     c_string_encoding="ascii",
-                    embedsignature=True,
-                    language_level=2,
+                    embedsignature=False,
+                    language_level=3,
                 ),
                 build_dir=os.path.join("build", "src"),
                 include_path=["include", package_include_dir, ffmpeg_include_dir],
